consolidate multi models
